+++ CHANGELOG.md
@@ -25,6 +25,14 @@ Notes
 ### New Features
 - **Add French (`fr`), German (`de`), Spanish (`es`) localization by generative AI**
   - Since I have no knowledge of languages other than Japanese and English, translations may contain strange translations. If you notice any translations that need correction, I would appreciate it if you could send feedback on what to change and how!
+- **Add the “Date and Time Display Format” option to the “Menu” section in General Settings**
+  - Changes how the date is displayed when copied. You can choose from “Absolute,” “Both,” or “Relative.”
+    - **Absolute (Default)**\
+      Example: 11/30/2025, 12:34PM
+    - **Both**\
+      Example: 11/30/2025, 12:34PM (5 minutes ago)
+    - **Relative**\
+      Example: 5 minutes ago
 
 ### Bug Fixes and Improvements
 - **Fix issue where the “Resume” button did not appear in the notification panel sent when launching the app while clipboard monitoring was paused**